+++ src/ProjectSetupServiceProvider.php
@@ -4,6 +4,9 @@ namespace Wijourdil\ProjectSetup;
 
 use Illuminate\Support\ServiceProvider;
 use Wijourdil\ProjectSetup\Console\Commands\SetupCommand;
+use Wijourdil\ProjectSetup\Services\PackageInstaller\ComposerPackageInstaller;
+use Wijourdil\ProjectSetup\Services\PackageInstaller\FakePackageInstaller;
+use Wijourdil\ProjectSetup\Services\PackageInstaller\PackageInstallerContract;
 
 class ProjectSetupServiceProvider extends ServiceProvider
 {
@@ -18,5 +21,15 @@ class ProjectSetupServiceProvider extends ServiceProvider
 
     public function register(): void
     {
+        if ($this->app->environment('testing')) {
+            $this->app->bind(PackageInstallerContract::class, FakePackageInstaller::class);
+        } else {
+            $this->app->bind(PackageInstallerContract::class, ComposerPackageInstaller::class);
+        }
+
+        $this->mergeConfigFrom(
+            __DIR__ . '/../config/project-setup.php',
+            'project-setup'
+        );
     }
 }