+++ src/Tasks/Contracts/Configurable.php
@@ -5,4 +5,6 @@ namespace Wijourdil\ProjectSetup\Tasks\Contracts;
 interface Configurable
 {
     public function configure(): void;
+
+    public function alreadyConfigured(): bool;
 }