+++ src/Tasks/Contracts/Executable.php
@@ -5,4 +5,6 @@ namespace Wijourdil\ProjectSetup\Tasks\Contracts;
 interface Executable
 {
     public function execute(): void;
+
+    public function alreadyExecuted(): bool;
 }