+++ src/Tasks/Contracts/Installable.php
@@ -5,4 +5,6 @@ namespace Wijourdil\ProjectSetup\Tasks\Contracts;
 interface Installable
 {
     public function install(): void;
+
+    public function alreadyInstalled(): bool;
 }