+++ src/Tasks/InstallAssertPackage.php
@@ -2,12 +2,17 @@
 
 namespace Wijourdil\ProjectSetup\Tasks;
 
-use Wijourdil\ProjectSetup\Tasks\Abstracts\ComposerPackageInstaller;
+use Wijourdil\ProjectSetup\Tasks\Abstracts\InstallComposerPackage;
 
-class InstallAssertPackage extends ComposerPackageInstaller
+class InstallAssertPackage extends InstallComposerPackage
 {
     protected function packageName(): string
     {
         return 'webmozart/assert';
+    }
+
+    protected function isDevDependency(): bool
+    {
+        return false;
     }
 }