+++ src/Tasks/InstallAssertPhpstanRulePackage.php
@@ -2,21 +2,23 @@
 
 namespace Wijourdil\ProjectSetup\Tasks;
 
-use Wijourdil\ProjectSetup\Tasks\Abstracts\ComposerDevPackageInstaller;
+use Wijourdil\ProjectSetup\Tasks\Abstracts\InstallComposerPackage;
 use Wijourdil\ProjectSetup\Tasks\Contracts\Configurable;
-use Wijourdil\ProjectSetup\Tasks\Contracts\Dependable;
-use Wijourdil\ProjectSetup\Tasks\Contracts\Outputable;
+use Wijourdil\ProjectSetup\Tasks\Contracts\HasDependencies;
 use Wijourdil\ProjectSetup\Tasks\Traits\CanWriteInFiles;
-use Wijourdil\ProjectSetup\Tasks\Traits\CanWriteToOutput;
 
-class InstallWebmozartAssertPhpstanRulePackage extends ComposerDevPackageInstaller implements Configurable, Dependable, Outputable
+class InstallAssertPhpstanRulePackage extends InstallComposerPackage implements Configurable, HasDependencies
 {
     use CanWriteInFiles;
-    use CanWriteToOutput;
 
     protected function packageName(): string
     {
         return 'phpstan/phpstan-webmozart-assert';
+    }
+
+    protected function isDevDependency(): bool
+    {
+        return true;
     }
 
     public function dependsOn(): array
@@ -34,9 +36,13 @@ class InstallWebmozartAssertPhpstanRulePackage extends ComposerDevPackageInstall
             'includes:',
             base_path('phpstan.neon')
         );
+    }
 
-        $this->info(
-            'Configuration file ' . base_path('phpstan.neon') . " had been updated for package {$this->packageName()}."
+    public function alreadyConfigured(): bool
+    {
+        return str_contains(
+            file_get_contents(base_path('phpstan.neon')),
+            "vendor/phpstan/phpstan-webmozart-assert/extension.neon"
         );
     }
 }