+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `wijourdil/project-setup` will be documented in this file.
 
+## 1.1.0 - 2022-02-26
+
+- added the detection of already ran tasks
+- added the possibility to force ignore / re-run already ran tasks
+
 ## 1.0.0 - 2022-02-25
 
 - initial release